+++ src/board/history.rs
@@ -14,7 +14,7 @@ use crate::defs::MAX_GAME_MOVES;
 
 #[derive(Clone)]
 pub struct History {
-    list: [GameState; MAX_GAME_MOVES as usize],
+    list: [GameState; MAX_GAME_MOVES],
     count: usize,
 }
 
@@ -22,7 +22,7 @@ impl History {
     // Create a new history array containing game states.
     pub fn new() -> Self {
         Self {
-            list: [GameState::new(); MAX_GAME_MOVES as usize],
+            list: [GameState::new(); MAX_GAME_MOVES],
             count: 0,
         }
     }
